+++ packages/tools/create-dlightjs/templates/dlight-vite-js/src/Header.view.js
@@ -1,19 +1,20 @@
 import { View } from "@dlightjs/dlight"
 import { headerWrap, navBtn, headerLogo } from "./style.module.css"
 
-class Header extends View {
+@View
+export default class Header {
   navList = [
     {
-      url: "",
+      url: "https://dlight-js.com/docs",
       navName: "Docs"
     },
     {
-      url: "https://www.baidu.com",
+      url: "https://dlight-js.com/examples",
       navName: "Examples"
     },
     {
-      url: "",
-      navName: "Tutorials"
+      url: "https://dlight-js.com/playground",
+      navName: "Playground"
     }
   ]
 
@@ -35,5 +36,3 @@ class Header extends View {
     }
   }
 }
-
-export default Header