+++ packages/tools/create-dlightjs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-dlightjs
 
+## 0.10.1
+
+### Patch Changes
+
+- feat: update templates
+
 ## 0.10.0
 
 ### Minor Changes