+++ packages/tools/transpiler/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @dlightjs/transpiler
 
+## 0.10.3
+
+### Patch Changes
+
+- Updated dependencies
+  - babel-preset-dlight@0.10.3
+
+## 0.10.2
+
+### Patch Changes
+
+- Updated dependencies
+  - babel-preset-dlight@0.10.2
+
 ## 0.10.1
 
 ### Patch Changes